Make widget editing pluggable
Use information on plugins store to render widget configuration
component.
