+++ graylog2-web-interface/src/components/widgets/configurations/index.js
@@ -1,0 +1,6 @@
+export { default as CountWidgetConfiguration } from './CountWidgetConfiguration';
+export { default as FieldChartWidgetConfiguration } from './FieldChartWidgetConfiguration';
+export { default as QueryConfiguration } from './QueryConfiguration';
+export { default as QuickValuesWidgetConfiguration } from './QuickValuesWidgetConfiguration';
+export { default as StackedChartWidgetConfiguration } from './StackedChartWidgetConfiguration';
+export { default as StatisticalCountWidgetConfiguration } from './StatisticalCountWidgetConfiguration';
