+++ graylog2-web-interface/src/components/widgets/types/index.js
@@ -5,6 +5,12 @@ import {
   NumericVisualization,
   QuickValuesVisualization,
   StackedGraphVisualization } from 'components/visualizations';
+import {
+  CountWidgetConfiguration,
+  FieldChartWidgetConfiguration,
+  QuickValuesWidgetConfiguration,
+  StackedChartWidgetConfiguration,
+  StatisticalCountWidgetConfiguration } from 'components/widgets/configurations';
 
 PluginStore.register(new PluginManifest({}, {
   widgets: [
@@ -13,18 +19,21 @@ PluginStore.register(new PluginManifest({}, {
       defaultHeight: 1,
       defaultWidth: 1,
       visualization: NumericVisualization,
+      configuration: CountWidgetConfiguration,
     },
     {
       type: 'STREAM_SEARCH_RESULT_COUNT',
       defaultHeight: 1,
       defaultWidth: 1,
       visualization: NumericVisualization,
+      configuration: CountWidgetConfiguration,
     },
     {
       type: 'STATS_COUNT',
       defaultHeight: 1,
       defaultWidth: 1,
       visualization: NumericVisualization,
+      configuration: StatisticalCountWidgetConfiguration,
     },
     {
       type: 'SEARCH_RESULT_CHART',
@@ -37,18 +46,21 @@ PluginStore.register(new PluginManifest({}, {
       defaultHeight: 1,
       defaultWidth: 2,
       visualization: QuickValuesVisualization,
+      configuration: QuickValuesWidgetConfiguration,
     },
     {
       type: 'FIELD_CHART',
       defaultHeight: 1,
       defaultWidth: 2,
       visualization: GraphVisualization,
+      configuration: FieldChartWidgetConfiguration,
     },
     {
       type: 'STACKED_CHART',
       defaultHeight: 1,
       defaultWidth: 2,
       visualization: StackedGraphVisualization,
+      configuration: StackedChartWidgetConfiguration,
     },
   ],
 }));
